python_oop_and_testing/requirements.txt
robotframework>=7.0,<8.0
robotframework-seleniumlibrary>=6.0,<7.0
selenium>=4.0,<5.0
pytest
pytest-mock
